requirements.txt
requests
pandas
pyarrow
fastparquet
matplotlib
backoff
seaborn
cliffs-delta
jupyter
scipy
huggingface_hub
python-dotenv